arxiv-visual-sprint/pyproject.toml
[project]
name = "arxiv-visual-sprint"
version = "0.1.0"
description = "45-minute arXiv visual abstract sprint tool with topic expansion"
readme = "SKILL.md"
requires-python = ">=3.9"
dependencies = [
    "google-genai>=0.3.0",
    "pillow>=10.0.0",
    "pypdf>=3.0.0",
    "requests>=2.32.0",
]

[tool.uv]
managed = true
